save rating test progress
Required modification to bookmark_rating model, specifying that rating must be an integer between 0 and 5 (inclusive).
Required modification to the save rating method, to make save! call (triggering errors on invalid input)
Only thing that is missing is copying the tests to users other than students.
ALSO, we may want to add more bounds tests
ALSO we need to run the rubocop thing and fix the test up