Attempted rework of self.varying_rubrics_report
The old self.varying_rubrics_report has been commented out, and changed to self.varying_rubrics_report_old

The new self.varying_rubrics_report starts with an empty dictionary, and has the keys added to it when a new one appears (the value is initialized to an empty array if the value is initially missing, that is what ||= does).
I added a note above line 141, since I'm not sure if it's interacting the way we want it to... but that WAS the logic of the existing function, so I don't want to touch it.

This change is untested, I have not ran the tests in a while, and do not want to go through the hassle of docker at the moment. PLEASE LET ME KNOW if these changes do not work.